cap nhat code moi

+++ ChatR/Controllers/MessageController.cs
@@ -1,5 +1,6 @@
 ﻿using ChatR.Data;
-using ChatR.DTOs;
+using ChatR.DTOs.Conversation;
+using ChatR.DTOs.Message;
 using ChatR.Helpers;
 using ChatR.Hubs;
 using ChatR.Models;
@@ -28,7 +29,7 @@ namespace ChatR.Controllers
         public async Task<IActionResult> SendMessage([FromBody] SendMessageDto dto, CancellationToken cancellationToken)
         {
             var senderIdString = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
-            if (!long.TryParse(senderIdString, out var senderId))
+            if (!int.TryParse(senderIdString, out var senderId))
                 return Unauthorized("Không lấy được userId từ token.");
 
             var result = await _messageService.SendMessageAsync(senderId, dto, cancellationToken);
@@ -46,7 +47,7 @@ namespace ChatR.Controllers
         public async Task<IActionResult> EditMessage([FromBody] EditMessageDto dto, CancellationToken cancellationToken)
         {
             var senderIdString = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
-            if (!long.TryParse(senderIdString, out var senderId))
+            if (!int.TryParse(senderIdString, out var senderId))
                 return Unauthorized();
 
             var result = await _messageService.EditMessageAsync(senderId, dto, cancellationToken);
@@ -57,7 +58,7 @@ namespace ChatR.Controllers
         public async Task<IActionResult> DeleteMessage(int messageId, CancellationToken cancellationToken)
         {
             var senderIdString = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
-            if (!long.TryParse(senderIdString, out var senderId))
+            if (!int.TryParse(senderIdString, out var senderId))
                 return Unauthorized();
 
             await _messageService.DeleteMessageAsync(senderId, messageId, cancellationToken);
