+++ ChatR/Models/Notification.cs
@@ -2,13 +2,12 @@
 {
     public class Notification
     {
-        public long NotificationId { get; set; }
+        public int NotificationId { get; set; }   // từ long → int
         public int Type { get; set; }
         public string Content { get; set; } = string.Empty;
         public int IsRead { get; set; } = 0;
         public DateTime CreatedAt { get; set; }
-        public long UserId { get; set; }
-
+        public int UserId { get; set; }           // từ long → int
         public User? User { get; set; }
     }
 }
